requirements.txt
torch
soundfile
numpy
pyannote.audio
scikit-learn
pydub